T#770: Extract telegram routes + polling to src/telegram/routes.ts
Telegram two-way chat module (polling, message handling, photo download,
2 route handlers, isTelegramAuthorized) moved from server.ts to
src/telegram/routes.ts. Polling starts on registerTelegramRoutes call.

server.ts: 12,335 → 11,963 lines (-372). No logic changes.
