+++ src/server.ts
@@ -95,6 +95,7 @@ import { registerGovernanceRoutes } from './governance/routes.ts';
 import { registerProwlRoutes } from './prowl/routes.ts';
 import { registerLibraryRoutes } from './library/routes.ts';
 import { registerRiskRoutes } from './risk/routes.ts';
+import { registerTelegramRoutes } from './telegram/routes.ts';
 import {
   initGuestTables,
   createGuest,
@@ -9604,28 +9605,7 @@ function isForgeAuthorized(c: any, options: { mode: 'read' | 'write' } = { mode:
 }
 
 // T#712 Telegram-cache read auth — DELIBERATELY SEPARATE from FORGE_BEAST_MODES per
-// Library #96 lever 1 (scope-for-post-compromise-damage). TG chat content has a
-// different sensitivity profile than Forge workout data: private Gorn-bot chats may
-// carry decisions, personal context, unreleased plans. Default narrow-by-default;
-// expansions require Tier-2-pair CLEAR per T#696 precedent (Bertus #887 flag 1).
-const TELEGRAM_READ_MODES: Record<string, 'read'> = {
-  sable: 'read', // routing-lane — DM→Prowl flow references TG context
-};
-function isTelegramAuthorized(c: any): boolean {
-  if (hasSessionAuth(c)) return true; // Gorn browser session — owner
-  // Per Gorn call 2026-04-24 23:42 BKK: any Beast with a configured TG bot
-  // can read the cache. Bearer-token (T#718) derives actor cryptographically;
-  // if actor matches a registered bot-beast, authorize.
-  const actor = (c.get as any)('actor') as string | undefined;
-  if (actor && telegramBots.some(b => b.beast === actor)) return true;
-  // Legacy back-compat: ?as= lookup in TELEGRAM_READ_MODES (preserved for
-  // pre-bearer-token callers until full rollout completes).
-  if (isTrustedRequest(c)) {
-    const as = (c.req.query('as') || '').toLowerCase();
-    return TELEGRAM_READ_MODES[as] === 'read';
-  }
-  return false;
-}
+// isTelegramAuthorized + TELEGRAM_READ_MODES moved to src/telegram/routes.ts (T#770)
 
 // GET /api/routine/logs — list logs
 app.get('/api/routine/logs', (c) => {
@@ -11722,360 +11702,8 @@ app.get('/api/search/status', async (c) => {
   });
 });
 
-// ============================================================================
-// Telegram Two-Way Chat (T#633)
-// ============================================================================
-// Multi-bot long-polling: each Beast can have its own Telegram bot.
-// Gorn sends messages/photos to a Beast's bot, they get forwarded as DMs.
-// Config: TELEGRAM_BOTS env var (JSON array) or legacy single-bot env vars.
-
-const TG_CHAT_ID = process.env.TELEGRAM_CHAT_ID || '';
-const TG_POLL_INTERVAL = 3000; // ms
-
-interface TelegramBot {
-  token: string;
-  beast: string;       // Beast name this bot belongs to
-  chatId: string;      // Gorn's chat ID (same for all bots)
-  offset: number;
-  lastMessageAt: string | null;
-  messageCount: number;
-  active: boolean;
-  timer: ReturnType<typeof setInterval> | null;
-  polling: boolean;    // Concurrent-poll guard. If a poll tick fires while a
-                       // previous poll is still running, skip — otherwise both
-                       // call getUpdates with the same offset and re-deliver
-                       // the same update twice (dup-delivery bug found 2026-04-17).
-}
-
-// Parse bot configs from env
-function parseTelegramBots(): TelegramBot[] {
-  const bots: TelegramBot[] = [];
-
-  // Try TELEGRAM_BOTS JSON array first: [{"token":"...","beast":"karo"},{"token":"...","beast":"sable"}]
-  const botsJson = process.env.TELEGRAM_BOTS;
-  if (botsJson) {
-    try {
-      const parsed = JSON.parse(botsJson);
-      for (const b of parsed) {
-        if (b.token && b.beast) {
-          bots.push({
-            token: b.token,
-            beast: b.beast,
-            chatId: b.chatId || TG_CHAT_ID,
-            offset: 0, lastMessageAt: null, messageCount: 0, active: false, timer: null, polling: false,
-          });
-        }
-      }
-    } catch (e) { console.error('[Telegram] Failed to parse TELEGRAM_BOTS:', e); }
-  }
-
-  // Fallback: legacy single-bot env vars
-  if (bots.length === 0) {
-    const token = process.env.TELEGRAM_BOT_TOKEN || '';
-    const beast = process.env.TELEGRAM_FORWARD_TO || 'karo';
-    if (token && TG_CHAT_ID) {
-      bots.push({
-        token, beast, chatId: TG_CHAT_ID,
-        offset: 0, lastMessageAt: null, messageCount: 0, active: false, timer: null, polling: false,
-      });
-    }
-  }
-
-  return bots;
-}
-
-const telegramBots = parseTelegramBots();
-
-async function tgApi(token: string, method: string, params: Record<string, string> = {}): Promise<any> {
-  const url = new URL(`https://api.telegram.org/bot${token}/${method}`);
-  for (const [k, v] of Object.entries(params)) url.searchParams.set(k, v);
-  const res = await fetch(url.toString());
-  return res.json();
-}
-
-async function tgSendReply(token: string, chatId: string, text: string): Promise<void> {
-  await tgApi(token, 'sendMessage', { chat_id: chatId, text });
-}
-
-async function handleTelegramMessage(bot: TelegramBot, msg: any): Promise<void> {
-  // Only accept messages from Gorn's chat.
-  // T#712 PII-containment gate — this check is the boundary that makes
-  // telegram_messages.raw_json safe to persist. Expanding this gate (group chats,
-  // multi-sender, Boro chat forwarded) requires threat-model re-review on the
-  // cached raw_json surface at the same time. Gate-to-storage coupling lives here.
-  if (String(msg.chat.id) !== bot.chatId) {
-    console.log(`[Telegram:${bot.beast}] Rejected: chat_id ${msg.chat.id} !== expected ${bot.chatId}`);
-    return;
-  }
-
-  // T#712 cache inbound message before notification build.
-  // Dual-shape mapper-lane per Library #98 canon-pending: silent-drop on malformed
-  // payload with console.warn noise-log (Hevy→Forge pattern mirror). Never rejects
-  // the notification path — just skips the cache write if the payload is shape-bad.
-  try {
-    const msgId = msg.message_id;
-    if (typeof msgId === 'number' && Number.isFinite(msgId)) {
-      // Strip known ephemeral TG-signed URL fields from raw_json before persist
-      // (Bertus #887 flag 3). TG file_paths regenerate on bot-token rotation and
-      // carry implicit auth; don't cache those.
-      const sanitized = JSON.parse(JSON.stringify(msg));
-      const stripEphemeral = (obj: any) => {
-        if (!obj || typeof obj !== 'object') return;
-        for (const k of Object.keys(obj)) {
-          if (k === 'file_path') delete obj[k]; // TG file download path (token-bearing)
-          else if (typeof obj[k] === 'object') stripEphemeral(obj[k]);
-        }
-      };
-      stripEphemeral(sanitized);
-      const rawJson = JSON.stringify(sanitized);
-      // INSERT OR IGNORE for TG retry-delivery dupes (Bertus #887 flag 4).
-      sqlite.prepare(
-        'INSERT OR IGNORE INTO telegram_messages (chat_id, id, from_id, text, caption, photo_file_id, date_unix, received_at, raw_json) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)'
-      ).run(
-        String(msg.chat.id),
-        msgId,
-        msg.from?.id ? String(msg.from.id) : null,
-        msg.text || null,
-        msg.caption || null,
-        // photo_file_id: TG file_ids regenerate on bot-token rotation; bytes not
-        // stored per v1 scope (photo bytes fetched separately in media branch below).
-        msg.photo && msg.photo.length > 0 ? (msg.photo[msg.photo.length - 1].file_id || null) : null,
-        msg.date || Math.floor(Date.now() / 1000),
-        Date.now(),
-        rawJson
-      );
-    } else {
-      console.warn(`[Telegram:${bot.beast} T#712] dropping cache — malformed message_id: ${JSON.stringify(msgId)}`);
-    }
-  } catch (cacheErr) {
-    console.warn(`[Telegram:${bot.beast} T#712] dropping cache — persist error:`, cacheErr);
-  }
-
-  try {
-    let notifyText: string;
-    let confirmText: string;
-
-    // Build reply-to context if this message is a reply.
-    // T#712 format: `(replying to TG#<id>: "<80-char-preview>...")`
-    // Parseable convention — `TG#` prefix, integer id, `:` delimiter, quoted preview.
-    // Caller can fetch full body via GET /api/telegram/message/:id when the 80-char
-    // preview truncates useful context.
-    let replyContext = '';
-    if (msg.reply_to_message) {
-      const replied = msg.reply_to_message;
-      const repliedId = typeof replied.message_id === 'number' ? replied.message_id : '?';
-      const repliedText = replied.text || replied.caption || '[media]';
-      const repliedPreview = repliedText.length > 80 ? repliedText.slice(0, 80) + '...' : repliedText;
-      replyContext = `(replying to TG#${repliedId}: "${repliedPreview}")\\n`;
-    }
-
-    if (msg.photo && msg.photo.length > 0) {
-      // Download and save photo so Beasts can view it
-      let photoUrl = '';
-      try {
-        const photo = msg.photo[msg.photo.length - 1];
-        const fileInfo = await tgApi(bot.token, 'getFile', { file_id: photo.file_id });
-        if (fileInfo.ok && fileInfo.result?.file_path) {
-          const filePath = fileInfo.result.file_path;
-          const imageRes = await fetch(`https://api.telegram.org/file/bot${bot.token}/${filePath}`);
-          if (imageRes.ok) {
-            const buffer = Buffer.from(await imageRes.arrayBuffer());
-            if (buffer.length <= 20 * 1024 * 1024) {
-              // Process with sharp if available
-              let processedBuffer = buffer;
-              let ext = '.' + (filePath.split('.').pop() || 'jpg');
-              try {
-                const sharp = require('sharp');
-                const metadata = await sharp(buffer).metadata();
-                if (metadata.width && metadata.width > 1920) {
-                  processedBuffer = await sharp(buffer).rotate().resize(1920, null, { withoutEnlargement: true }).jpeg({ quality: 95 }).withMetadata({ orientation: undefined }).toBuffer();
-                  ext = '.jpg';
-                } else {
-                  processedBuffer = await sharp(buffer).rotate().withMetadata({ orientation: undefined }).toBuffer();
-                }
-              } catch { /* sharp not available */ }
-
-              if (!fs.existsSync(UPLOADS_DIR)) fs.mkdirSync(UPLOADS_DIR, { recursive: true });
-              const filename = `telegram_${crypto.randomUUID()}${ext}`;
-              fs.writeFileSync(path.join(UPLOADS_DIR, filename), processedBuffer);
-              try {
-                sqlite.prepare('INSERT INTO files (filename, original_name, mime_type, size_bytes, uploaded_by, context, created_at) VALUES (?, ?, ?, ?, ?, ?, ?)').run(filename, `telegram_photo${ext}`, ext === '.jpg' ? 'image/jpeg' : 'image/png', processedBuffer.length, 'gorn', 'telegram', Date.now());
-              } catch { /* files table may not have all columns */ }
-              photoUrl = `https://denbook.online/api/f/${filename}`;
-              console.log(`[Telegram:${bot.beast}] Photo saved: ${filename}`);
-            }
-          }
-        }
-      } catch (e) { console.error(`[Telegram:${bot.beast}] Photo download error:`, e); }
-
-      // T#712: replyContext threaded to photo branch (walk-all-write-paths per @pip discipline)
-      const caption = msg.caption || '';
-      if (photoUrl) {
-        notifyText = caption
-          ? `[Telegram from Gorn] ${replyContext}${caption}\\n\\nPhoto: ${photoUrl}`
-          : `[Telegram from Gorn] ${replyContext}Photo: ${photoUrl}`;
-      } else {
-        notifyText = caption
-          ? `[Telegram from Gorn] ${replyContext}Photo: ${caption}`
-          : `[Telegram from Gorn] ${replyContext}Photo received (download failed)`;
-      }
-      confirmText = `✓ Notified ${bot.beast}`;
-
-    } else if (msg.text) {
-      notifyText = `[Telegram from Gorn] ${replyContext}${msg.text}`;
-      confirmText = `✓ Notified ${bot.beast}`;
-
-    } else if (msg.document) {
-      // T#712: replyContext threaded to document branch
-      const docName = msg.document.file_name || 'unknown';
-      notifyText = `[Telegram from Gorn] ${replyContext}Document: ${docName}${msg.caption ? ' — ' + msg.caption : ''}`;
-      confirmText = `✓ Notified ${bot.beast}`;
-
-    } else if (msg.voice) {
-      // T#712: replyContext threaded to voice branch
-      notifyText = `[Telegram from Gorn] ${replyContext}Voice message`;
-      confirmText = `✓ Notified ${bot.beast}`;
-
-    } else if (msg.sticker) {
-      // T#712: replyContext threaded to sticker branch
-      const emoji = msg.sticker.emoji || '';
-      notifyText = `[Telegram from Gorn] ${replyContext}Sticker ${emoji}`;
-      confirmText = `✓ Notified ${bot.beast}`;
-
-    } else {
-      notifyText = `[Telegram from Gorn] ${replyContext}Message received`;
-      confirmText = `✓ Notified ${bot.beast}`;
-    }
-
-    // Use the TG message's send time (msg.date is unix seconds) so the Beast
-    // sees when Gorn actually SENT the message, not when polling caught it.
-    // Important on bad connectivity (trains, tunnels) where the gap can be
-    // meaningful. Falls back to server receive time if msg.date missing.
-    // The timestamp itself is stamped by enqueueNotification — we just pass
-    // the event time through via opts.sentAt.
-    const msgTime = msg.date ? new Date(msg.date * 1000) : new Date();
-
-    // Send tmux notification to the Beast — they reply via Telegram
-    const notification = `${notifyText}\\n\\nReply via Telegram to respond to Gorn.`;
-    enqueueNotification(bot.beast, notification, { sentAt: msgTime });
-
-    console.log(`[Telegram:${bot.beast}] Notified: ${notifyText.slice(0, 80)}`);
-    bot.messageCount++;
-    bot.lastMessageAt = new Date().toISOString();
-
-  } catch (err) {
-    console.error(`[Telegram:${bot.beast}] Error handling message:`, err);
-  }
-}
-
-async function pollTelegramBot(bot: TelegramBot): Promise<void> {
-  // Concurrent-poll guard: if a previous poll hasn't returned yet (common when
-  // long-poll timeout + async handleTelegramMessage exceed TG_POLL_INTERVAL),
-  // skip this tick. Without this, two polls race on the same bot.offset and
-  // re-deliver the same update_id twice.
-  if (bot.polling) return;
-  bot.polling = true;
-  try {
-    const data = await tgApi(bot.token, 'getUpdates', {
-      offset: String(bot.offset),
-      timeout: '3',
-      allowed_updates: JSON.stringify(['message']),
-    });
-
-    if (data.ok && data.result?.length) {
-      for (const update of data.result) {
-        bot.offset = update.update_id + 1;
-        const msg = update.message;
-        if (msg) {
-          console.log(`[Telegram:${bot.beast}] Update ${update.update_id}: chat_id=${msg.chat?.id} from=${msg.from?.username || msg.from?.id} text=${(msg.text || '[non-text]').slice(0, 50)}`);
-          await handleTelegramMessage(bot, msg);
-        }
-      }
-    }
-  } catch (err) {
-    console.error(`[Telegram:${bot.beast}] Poll error:`, err);
-  } finally {
-    bot.polling = false;
-  }
-}
-
-function startTelegramPolling(): void {
-  if (telegramBots.length === 0) {
-    console.log('[Telegram] No bots configured — polling disabled');
-    return;
-  }
-
-  for (const bot of telegramBots) {
-    bot.active = true;
-    // Stagger initial polls to avoid hitting Telegram rate limits
-    const delay = telegramBots.indexOf(bot) * 1000;
-    setTimeout(() => {
-      pollTelegramBot(bot).then(() => {
-        bot.timer = setInterval(() => pollTelegramBot(bot), TG_POLL_INTERVAL);
-        console.log(`[Telegram:${bot.beast}] Polling started (every ${TG_POLL_INTERVAL / 1000}s)`);
-      });
-    }, delay);
-  }
-}
-
-// GET /api/telegram/status — polling status (owner only)
-app.get('/api/telegram/status', (c) => {
-  if (!hasSessionAuth(c) && !isTrustedRequest(c)) return c.json({ error: 'Auth required' }, 403);
-
-  return c.json({
-    bots: telegramBots.map(b => ({
-      beast: b.beast,
-      polling: b.active,
-      chat_id: b.chatId ? `${b.chatId.slice(0, 4)}****` : null,
-      last_message_at: b.lastMessageAt,
-      message_count: b.messageCount,
-    })),
-    poll_interval_ms: TG_POLL_INTERVAL,
-    total_bots: telegramBots.length,
-  });
-});
-
-// T#712: GET /api/telegram/message/:id — fetch cached inbound TG message body.
-// Use when the 80-char replyContext preview in a Beast notification is insufficient.
-// Narrow-by-default auth (TELEGRAM_READ_MODES: Gorn session + Sable only).
-// `:id` is the TG message_id (integer); resolves against single-chat cache since
-// the chat-gate is Gorn-only upstream. Composite PK in the table allows this to
-// expand cleanly if the chat-gate ever opens.
-app.get('/api/telegram/message/:id', (c) => {
-  if (!isTelegramAuthorized(c)) return c.json({ error: 'Telegram cache is private' }, 403);
-  const idParam = c.req.param('id');
-  const msgId = parseInt(idParam, 10);
-  if (!Number.isFinite(msgId) || String(msgId) !== idParam) {
-    return c.json({ error: 'id must be an integer' }, 400);
-  }
-  // T#712 Bertus #890 harden: query uses composite PK scoped to configured
-  // bot chat_ids rather than WHERE id=? alone. Future-proof against chat-gate
-  // expansion without relying on upstream-gate assumption. Single-chat universe
-  // today (all bots share Gorn's chatId) so IN clause is 1-wide.
-  const validChatIds = telegramBots.map(b => b.chatId).filter(Boolean);
-  if (validChatIds.length === 0) return c.json({ error: 'no telegram bots configured' }, 503);
-  const placeholders = validChatIds.map(() => '?').join(',');
-  const row = sqlite.prepare(
-    `SELECT chat_id, id, from_id, text, caption, photo_file_id, date_unix, received_at, raw_json FROM telegram_messages WHERE chat_id IN (${placeholders}) AND id = ? LIMIT 1`
-  ).get(...validChatIds, msgId) as any;
-  if (!row) return c.json({ error: 'message not found' }, 404);
-  let raw: any = null;
-  try { raw = JSON.parse(row.raw_json); } catch { /* leave null on parse fail */ }
-  return c.json({
-    chat_id: row.chat_id,
-    id: row.id,
-    from_id: row.from_id,
-    text: row.text,
-    caption: row.caption,
-    photo_file_id: row.photo_file_id,
-    date_unix: row.date_unix,
-    received_at: row.received_at,
-    raw: raw,
-  });
-});
-
-// Start polling on server boot
-startTelegramPolling();
+// Telegram routes + polling — extracted to src/telegram/routes.ts (T#770)
+registerTelegramRoutes(app, sqlite, { hasSessionAuth, isTrustedRequest, uploadsDir: UPLOADS_DIR });
 
 // ============================================================================
 // OpenAPI Schema + Swagger UI (Spec #55 Phase 1)
